+++ packages/xchain-aggregator/__tests__/aggregator.test.ts
@@ -15,6 +15,7 @@ import mockMayaMidgardApi from '../__mocks__/mayachain/midgard/api'
 import mockThorMidgardApi from '../__mocks__/thorchain/midgard/api'
 import mockMayanodeApi from '../__mocks__/thorchain/thornode/api'
 import { Aggregator } from '../src'
+import { SupportedProtocols } from '../src/const'
 
 describe('Aggregator', () => {
   let aggregator: Aggregator
@@ -35,6 +36,36 @@ describe('Aggregator', () => {
     mockThorMidgardApi.restore()
     mockMayanodeApi.restore()
     mockMayaMidgardApi.restore()
+  })
+
+  it('Should init with no parameters', () => {
+    expect(() => new Aggregator()).not.toThrowError()
+  })
+
+  it('Should init with all protocols enabled', () => {
+    const aggregator = new Aggregator()
+    expect(aggregator.getConfiguration().protocols.length).toBe(SupportedProtocols.length)
+  })
+
+  it('Should init with custom protocols enabled', () => {
+    const aggregator = new Aggregator({ protocols: ['Thorchain'] })
+    expect(aggregator.getConfiguration().protocols.length).toBe(1)
+  })
+
+  it('Should throw error if no protocols enabled', () => {
+    expect(() => new Aggregator({ protocols: [] })).toThrowError('No protocols enabled')
+  })
+
+  it('Should throw error if basis points lower than 0', () => {
+    expect(() => new Aggregator({ affiliate: { basisPoints: -1, affiliates: {} } })).toThrowError(
+      'Invalid affiliate basis point due to it is out of bound. It must be between [0 - 10000]',
+    )
+  })
+
+  it('Should throw error if basis points greater than 10000', () => {
+    expect(() => new Aggregator({ affiliate: { basisPoints: 10001, affiliates: {} } })).toThrowError(
+      'Invalid affiliate basis point due to it is out of bound. It must be between [0 - 10000]',
+    )
   })
 
   it('Should find swap with greatest expected amount', async () => {